Move MakeFeature to a separate module
